requirements.txt
tqdm
imageio[ffmpeg]
torch
torchvision
torchaudio
jaxtyping
tyro
git+https://github.com/NVlabs/nvdiffrast.git
kiui
opencv-python
trimesh
scipy
huggingface-hub[torch]>=0.22
torchsnooper
matplotlib==3.8.4
gpustat
plyfile
alphashape
scikit-image
piqa
git+https://github.com/cubao/concave_hull.git
einops
roma
alphashape
pandas
geopandas
scikit-learn
shapely
viser
tensorboard
tensorboardX
gradio
rtree
pillow
git+https://github.com/facebookresearch/segment-anything.git
evo